% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{u.predict}
\alias{u.predict}
\title{Predicting classes from a univariate decision tree}
\usage{
u.predict(t, X, ...)
}
\arguments{
\item{t}{is a tree from fitted by \code{unitree}}

\item{X}{is a dataframe with the same columns fitting 't'.}
}
\value{
a vector of class labels for each row from 'x'
}
\description{
This function predict classes from a dataframe 'x' to classes 'y' based on a \code{unitree} object.
}
\details{
This function provides class predictions from a dataframe 'x' based on a \code{unitree} object. This prediction function traverses the tree until reaching a leaf node where the label is returned.
}
\examples{
Y  <- iris[,5]
X  <- iris[,1:4]
dt <- unitree(X, Y)
p  <- u.predict(dt, X)
table(pred = p, actu = Y)
}
\author{
David Rodriguez
}
\seealso{
\code{unitree, u.predict}
}

