% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.R
\name{uniforest}
\alias{uniforest}
\title{A random forest classifier}
\usage{
uniforest(X, Y, n, grow = TRUE, splitter = Branch, thresh = KTile, ...)
}
\arguments{
\item{X}{dataframe of real-values (does not contain missing values).}

\item{Y}{a vector of class set as factors.}

\item{n}{the number of trees to make.}

\item{grow}{If \code{TRUE} then either bagging or random selection of columns of \code{X}.}

\item{splitter}{Either: Split or Branch. If split, then traditional univariate decision tree. If branch, then random forest, and potential sum of attributes.}

\item{Thresh}{Either: Gauss, KTile, Uniform, Brute. Approximation splitting techniques.}

\item{Pure}{Either: Gini, Info, Twoing. Funcitons to calculate impurity at a given node.}
}
\value{
a \code{uniforest} object consisting of \code{unitree} objects consisting of nodes.
}
\description{
A function to construct a univariate and sum of attributes random forest.
}
\details{
This function fits a random forest with the above options. It is therefore, a potentially, exhaustive approach to fitting and selecting a random forest classifier
}
\examples{
X <- iris[,1:4]
Y <- iris[,5]
dt <- uniforest(X, Y, 10)
}
\author{
David Rodriguez
}
\seealso{
\code{rf.predict}
}

