% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{unitree}
\alias{unitree}
\title{A decision tree and random forest classifier}
\usage{
unitree(X, Y, Thresh = Gauss, Pure = Info, is.forest = FALSE,
  splitter = Split, ...)
}
\arguments{
\item{X}{dataframe of real-values (does not contain missing values).}

\item{Y}{a vector of class set as factors.}

\item{Thresh}{Either: Gauss, KTile, Uniform, Brute. Approximation splitting techniques.}

\item{Pure}{Either: Gini, Info, Twoing. Funcitons to calculate impurity at a given node.}

\item{is.forest}{If \code{TRUE} then either bagging or random selection of columns of \code{X}.}

\item{splitter}{Either: Split or Branch. If split, then traditional univariate decision tree. If branch, then random forest, and potential sum of attributes.}
}
\value{
a \code{tree} object consisting of nodes.
}
\description{
A function to construct a univariate decision tree classifier or random forest.
}
\details{
This function fits a decision tree with the above options. It is therefore, a potentially, exhaustive approach to fitting and selecting a decision tree.
}
\examples{
X <- iris[,1:4]
Y <- iris[,5]
dt <- unitree(X, Y)
}
\author{
David Rodriguez
}
\seealso{
\code{u.predict, force.graph}
}

